pyproject.toml
[project]
name = "encontraplaca"
version = "0.1.0"
description = ""
authors = [
    {name = "douglas",email = "douglaspaz955@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "ultralytics (>=8.3.198,<9.0.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false